// interfaces/ExerciseCategoriesEnum.ts
export enum ExerciseCategories {
    PLACEHOLDER = "- Bitte Auswählen -",
    SQUAT = "Squat",
    BENCH = "Bench",
    DEADLIFT = "Deadlift",
    OVERHEADPRESS = "Overheadpress",
    CHEST = "Chest",
    BACK = "Back",
    SHOULDER = "Shoulder",
    TRICEPS = "Triceps",
    BICEPS = "Biceps",
    LEGS = "Legs"
  }